refactor : component 위치 통일화
