+++ app/page.tsx
@@ -1,11 +1,11 @@
 import { getCategories } from '@/lib/api/category';
-import HeaderSection from '@/app/_components/HeaderSection';
 import PostListSuspense from '@/components/features/blog/PostListSuspense';
 import { Suspense } from 'react';
 import CategorySectionClient from '@/components/features/category/CategorySection.client';
 import CategorySectionSkeleton from '@/components/features/category/CategorySectionSkeleton';
 import PostListSkeleton from '@/components/features/blog/PostListSkeleton';
 import { getPostList } from '@/lib/api/post';
+import BlogMainHeaderSection from '@/components/features/blog/main/BlogMainHeaderSection';
 
 interface HomeProps {
   searchParams: Promise<{ category?: string; sort?: string; keyword?: string }>;
@@ -20,7 +20,7 @@ export default async function Home({ searchParams }: HomeProps) {
     <div className="container py-8">
       {/* 메인페이지 헤더 섹션, 추후 검색 추가 */}
       <div className="mb-8">
-        <HeaderSection />
+        <BlogMainHeaderSection />
       </div>
       <div className="grid grid-cols-1 gap-6 md:grid-cols-[200px_1fr]">
         {/* 좌측 사이드바 */}
